# main/swagger.yaml
basePath: /api/config
consumes:
- application/json
definitions:
  ChangeLog:
    description: ChangeLog defines the structure for an API config-log
    properties:
      AgentInfo:
        description: the Agent Info for the change log entry
        maxLength: 20
        type: string
      ApplicationName:
        description: the Application Name for the change log entry
        maxLength: 20
        type: string
      CommitHash:
        description: the Commit Hash for the change log entry
        maxLength: 100
        type: string
        uniqueItems: true
      ConfigEntryID:
        description: the id for the change log entry - auto genrated
        format: uint64
        minimum: 1
        pattern: '[0-9]+'
        type: integer
        uniqueItems: true
        x-go-name: ID
      CreatedAt:
        description: the Creation Date for the change log entry - auto genrated
        format: date-time
        type: string
        uniqueItems: true
      EnvironmentName:
        description: the Environment Name for the change log entry
        maxLength: 10
        type: string
      Message:
        description: the Message for the change log entry
        maxLength: 255
        type: string
      ReleaseInfo:
        description: the Release Info for the change log entry
        maxLength: 30
        type: string
      ServiceTeamName:
        description: the Service Team Name for the change log entry
        maxLength: 20
        type: string
      UpdatedAt:
        description: the Modification Date for the change log entry - auto genrated
        format: date-time
        type: string
      Username:
        description: the Username for the change log entry (Execution user)
        maxLength: 20
        type: string
    required:
    - ServiceTeamName
    - ApplicationName
    - Username
    - EnvironmentName
    - CommitHash
    - ReleaseInfo
    - AgentInfo
    - Message
    type: object
    x-go-package: github.com/nitinda/microservice-change-log/api/models
  GenericError:
    description: GenericError is a Generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/nitinda/microservice-change-log/api/docs
  TeamInfo:
    description: TeamInfo struct abstract of Team data table in database
    properties:
      ClientSecret:
        description: the Client Secret to generate new session token
        example: '''$2a$10$m/22n7xmifpi1/rpsIzsIuY7.9walzEKloGCJF2ZpV.AElO83f2du'''
        maxLength: 80
        type: string
        uniqueItems: true
      CreatedAt:
        format: date-time
        type: string
      TeamID:
        format: uint32
        type: integer
      TeamName:
        description: the Service Team Name to generate new session token
        example: sales
        maxLength: 20
        type: string
        uniqueItems: true
      UpdatedAt:
        format: date-time
        type: string
    required:
    - TeamName
    - ClientSecret
    type: object
    x-go-package: github.com/nitinda/microservice-change-log/api/models
host: localhost
info:
  contact:
    email: nitindas@gmail.com
    name: Nitin Das
  description: |-
    the purpose of this application is to provide an application
    that will be used to record config chanegs applied to services.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: there are no TOS at this moment.
  title: changelog.
  version: 1.0.0
paths:
  /api/config:
    post:
      consumes:
      - application/json
      description: Create new config log entry
      operationId: createChangeLog
      parameters:
      - in: body
        name: Body
        schema:
          properties:
            AgentInfo:
              description: the Agent Info for the change log entry
              example: jenkins | gitlab
              maxLength: 20
              type: string
            ApplicationName:
              description: the Application Name for the change log entry
              example: ngs
              maxLength: 20
              type: string
            CommitHash:
              description: the Commit Hash for the change log entry
              maxLength: 100
              type: string
              uniqueItems: true
            EnvironmentName:
              description: the Environment Name for the change log entry
              example: dev | test | prod
              maxLength: 10
              type: string
            Message:
              description: the Message for the change log entry
              example: This is dummy release
              maxLength: 255
              type: string
            ReleaseInfo:
              description: the Release Info for the change log entry
              example: Release number 1.2.3
              maxLength: 30
              type: string
            ServiceTeamName:
              description: the Service Team Name for the change log entry
              example: sales
              maxLength: 20
              type: string
            Username:
              description: the Username for the change log entry (Execution user)
              example: John Wick
              maxLength: 20
              type: string
          required:
          - ServiceTeamName
          - ApplicationName
          - Username
          - EnvironmentName
          - CommitHash
          - ReleaseInfo
          - AgentInfo
          - Message
          type: object
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/changelogResponse'
        "401":
          $ref: '#/responses/changelogErrorResponse'
      schemes:
      - http
      security:
      - SecurityDefinitions: []
      - bearer: []
      - name:
        - Authorization
      - in:
        - body
      tags:
      - changelog
  /api/token:
    post:
      consumes:
      - application/json
      description: Create new session token entry
      operationId: createSessionToken
      parameters:
      - in: body
        name: Body
        schema:
          properties:
            ClientSecret:
              description: the Client Secret to generate new session token
              example: '''$2a$10$m/22n7xmifpi1/rpsIzsIuY7.9walzEKloGCJF2ZpV.AElO83f2du'''
              maxLength: 80
              type: string
              uniqueItems: true
            TeamName:
              description: the Service Team Name to generate new session token
              example: sales
              maxLength: 20
              type: string
              uniqueItems: true
          required:
          - TeamName
          - ClientSecret
          type: object
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/createSessionTokenResponse'
        "403":
          $ref: '#/responses/createSessionTokenErrorResponse'
      schemes:
      - http
      tags:
      - changelog
produces:
- application/json
responses:
  changelogErrorResponse:
    description: Generic error message returned as a string
    schema:
      properties:
        unauthorized:
          example: Token is expired or signature is invalid
          type: string
          x-go-name: Message
      type: object
  changelogResponse:
    description: Data structure representing a single changelog
    schema:
      $ref: '#/definitions/ChangeLog'
  createSessionTokenErrorResponse:
    description: A ValidationError is an error that is used when the required input fails validation.
    schema:
      properties:
        unauthorized:
          example: Access Denied
          type: string
          x-go-name: Message
      type: object
  createSessionTokenResponse:
    description: Data structure representing create session token api response
    schema:
      properties:
        access_token:
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MDM3NTc2NzgsInRlYW1JRCI6MX0.uCRZnm6Fz6nI5_dV9lzRKejF3M5HB_MD5gRYuHG8sfVfPVmdNH
          type: string
          x-go-name: Code
      type: object
  genericError:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
schemes:
- http
swagger: "2.0"